# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MiniStay API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/rooms": {
            "post": {
                "tags": [
                    "Admin - Rooms"
                ],
                "summary": "Create a new room",
                "operationId": "ebee0ee5660d504a8725690b769205a4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price_per_day"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "VIP Suite"
                                    },
                                    "price_per_day": {
                                        "type": "integer",
                                        "example": 350000
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Kamar luas dengan fasilitas lengkap"
                                    },
                                    "facilities": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "AC, TV, WiFi"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Room created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "adminAuth": []
                    }
                ]
            }
        },
        "/api/admin/rooms/{id}": {
            "put": {
                "tags": [
                    "Admin - Rooms"
                ],
                "summary": "Update room data",
                "operationId": "bf445be7b83b488608cda403f67cfa0e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Room ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Superior Room"
                                    },
                                    "price_per_day": {
                                        "type": "integer",
                                        "example": 250000
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Kamar nyaman untuk 2 orang"
                                    },
                                    "facilities": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Room updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "adminAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Rooms"
                ],
                "summary": "Delete a room",
                "operationId": "7788c9cbe476de80dba94397c381a034",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Room ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Room deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "adminAuth": []
                    }
                ]
            }
        },
        "/api/admin/rooms/{id}/upload-photo": {
            "post": {
                "tags": [
                    "Admin - Rooms"
                ],
                "summary": "Upload room photo",
                "operationId": "e7d2067222e9e0f14c36de9524078c21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Room ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "photo"
                                ],
                                "properties": {
                                    "photo": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "is_360": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Photo uploaded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Photo uploaded"
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "https://supabase.storage/rooms/photo123.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "adminAuth": []
                    }
                ]
            }
        },
        "/api/admin/rooms/{id}/availability": {
            "post": {
                "tags": [
                    "Admin - Rooms"
                ],
                "summary": "Update room availability for a specific date",
                "operationId": "9c47a7e82098cb0c06ab1db43761a551",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Room ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date",
                                    "status"
                                ],
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "example": "2025-01-10"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "available",
                                            "booked"
                                        ],
                                        "example": "booked"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Availability updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Availability updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "adminAuth": []
                    }
                ]
            }
        },
        "/api/admin/rooms/{id}/calendar": {
            "get": {
                "tags": [
                    "Admin - Rooms"
                ],
                "summary": "Get full availability calendar for a room",
                "operationId": "b54c781b99896dcc9fab0b13be6e8079",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Room ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Availability calendar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RoomAvailability"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "adminAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Room": {
                "required": [
                    "name",
                    "price_per_day"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "price_per_day": {
                        "type": "integer"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "facilities": {
                        "type": "object",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "RoomAvailability": {
                "required": [
                    "room_id",
                    "date",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "room_id": {
                        "type": "integer"
                    },
                    "date": {
                        "type": "string",
                        "format": "date"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "available",
                            "booked"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}